+++ dcm4chee/dcm4chee-arr/trunk/dcm4chee-arr-entities/src/main/java/org/dcm4chee/arr/entities/AuditRecord.java
@@ -77,7 +77,7 @@ public class AuditRecord implements Serializable {
 
     private static final Logger log = Logger.getLogger(AuditRecord.class);
 
-    private int pk;
+    private long pk;
 
     private Code eventID;
 
@@ -111,11 +111,11 @@ public class AuditRecord implements Serializable {
             @Parameter(name = "max_lo", value = "100"),
             @Parameter(name = "sequence", value = "audit_record_pk_seq") })
     @Column(name = "pk")
-    public int getPk() {
+    public long getPk() {
         return pk;
     }
 
-    public void setPk(int pk) {
+    public void setPk(long pk) {
         this.pk = pk;
     }
